requirements.txt
numpy
Pillow
reedsolo 